+++ cuda_src/CUDA_AUX.h
@@ -18,7 +18,7 @@ inline void gpuAssert(cudaError_t code, const char* file, int line) {
 }
 
 __global__ void
-GPU_rmse(long const* __restrict__ test_row, long const* __restrict__ test_col, float const* __restrict__ test_val,
+GPU_rmse(unsigned const* __restrict__ test_row, unsigned const* __restrict__ test_col, float const* __restrict__ test_val,
          float* __restrict__ pred_v, float* __restrict__ rmse, float const* __restrict__ W,
          float const* __restrict__ H, int m, int k, int rows, int cols, bool ifALS);
